# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "certificate-manage",
        "version": "1.0",
        "description": "証明書管理サービス"
    },
    "servers": [
        {
            "url": "/certificate-manage"
        }
    ],
    "paths": {
        "/": {
            "post": {
                "tags": [
                    "証明書保存"
                ],
                "description": "ルート証明書、サーバ証明書、クライアント証明書を受け、保存します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "sessionId",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "loginid",
                                "name": "test-user",
                                "mobilePhone": "0311112222",
                                "auth": {
                                    "add": true,
                                    "update": true,
                                    "delete": true
                                },
                                "lastLoginAt": "2020-01-01T00:00:00.000+0900",
                                "attributes": {},
                                "roles": [
                                    {
                                        "_value": 1,
                                        "_ver": 1
                                    }
                                ],
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "証明書保存",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "証明書保存のリクエストボディ",
                                "properties": {
                                    "certType": {
                                        "type": "string",
                                        "description": "証明書タイプ",
                                        "example": "root"
                                    },
                                    "serialNo": {
                                        "type": "string",
                                        "description": "シリアル番号",
                                        "example": "XXXXXXXX"
                                    },
                                    "fingerPrint": {
                                        "type": "string",
                                        "description": "フィンガープリント",
                                        "example": "XX:XX:XX:XX:XX:XX:XX:XX"
                                    },
                                    "certificate": {
                                        "type": "string",
                                        "description": "証明書",
                                        "example": "-----BEGIN RSA PRIVATE KEY-----\r\nMIGrAgEAAiEAvnrd8LBnzAGxCW+i7KtVQSiTsssMtbwcs5styeKsn2kCAwEAAQIh\r\nAKBF8glb5Xqa0cQG0ygg4hIFdipmvEJhiCuhX93krDCBAhEA51bAM0gFPvxyk9Xe\r\nioIOBQIRANLJEv4Xw7MwT7EEEARL5RUCEBa8bu1bUbCsDPK8nT+NoqUCEQCIzFCU\r\nMY4j7BW8N3vBnhPlAhBgs4tSfe6RbpertixmCygk\r\n-----END RSA PRIVATE KEY-----\r\n-----BEGIN CERTIFICATE-----\r\nMIGrAgEAAiEAvnrd8LBnzAGxCW+i7KtVQSiTsssMtbwcs5styeKsn2kCAwEAAQIh\r\nAKBF8glb5Xqa0cQG0ygg4hIFdipmvEJhiCuhX93krDCBAhEA51bAM0gFPvxyk9Xe\r\nioIOBQIRANLJEv4Xw7MwT7EEEARL5RUCEBa8bu1bUbCsDPK8nT+NoqUCEQCIzFCU\r\nMY4j7BW8N3vBnhPlAhBgs4tSfe6RbpertixmCygk\r\n-----END CERTIFICATE-----"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "証明書の保存に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "レスポンス",
                                    "properties": {
                                        "certType": {
                                            "type": "string",
                                            "example": "root"
                                        },
                                        "subject": {
                                            "type": "object",
                                            "properties": {
                                                "C": {
                                                    "type": "string",
                                                    "example": "JP"
                                                },
                                                "ST": {
                                                    "type": "string",
                                                    "example": "Tokyo"
                                                },
                                                "L": {
                                                    "type": "string",
                                                    "example": "Minato-ku"
                                                },
                                                "O": {
                                                    "type": "string",
                                                    "example": "aaa Corporation"
                                                },
                                                "OU": {
                                                    "type": "string",
                                                    "example": "PXR"
                                                },
                                                "CN": {
                                                    "type": "string",
                                                    "example": "*.---.co.jp"
                                                }
                                            }
                                        },
                                        "serialNo": {
                                            "type": "string",
                                            "example": "2CF1B929DB73BD43DB8206408D0C3073EC2B419B"
                                        },
                                        "fingerPrint": {
                                            "type": "string",
                                            "example": "09:FF:AA:8C:44:B4:90:DB:1A:E4:84:59:3E:93:42:1D:FA:5B:F8:13"
                                        },
                                        "validPeriodStart": {
                                            "type": "string",
                                            "example": "2020-03-26T21:37:02.000+0900"
                                        },
                                        "validPeriodEnd": {
                                            "type": "string",
                                            "example": "2120-03-02T21:37:02.000+0900"
                                        },
                                        "certificate": {
                                            "type": "string",
                                            "example": "-----BEGIN RSA PRIVATE KEY-----\r\nMIIEpQIBAAKCAQEAsDwJjxkOwSWufafRvREUO3MLeyLuOj0pXh2cpw5VNS3hiflv\r\nWl8cMlh5h4pp4lAI4u+ML+7YK8gc7Blk6y6GYDVFrR4KggpOPIi78Z7VeuD6w0Wu\r\n7PS5sclj7g4Lk1PxnS4G0uIOAm+mvUcUrNZFN4it5is3smSG9Mp6YLh+5/VMHT4/\r\nH86aqd3dVgIaTz8kcnbOAGBBjjn1YsZuf7UsbPf3ObOG3+NpE/Rk1b/s3nRkNFdZ\r\nx4K4D+9dujIJfZ6nNn61RL3LqUm6eAt6qDr/1quiNiNLND1rk0X+ojJ/LaHvBMD3\r\nxxSm5IMRMhEbENchcJLAVsnjDuEjhlMa98dnFQIDAQABAoIBAQChHoJuvS3sD5Mz\r\ndhYekFq67zbz4L+Dpx8G2HO7lzL3Z3bnOZBAQuv4a0LQVA7xxNIbB35W5Dnl7RHk\r\nGGNXUDWyB4In/2sD8IAOf8Oz5s6thLo9hJukLykGDWS640RZ8UjXtpiue9TrDUCX\r\nk0Cns02cs+yonaJSCe4IJoUzn/kF/ormTd3AbMNSCS/fGCVEgDE+0Gzfw4eFhVXP\r\nhwuR8rVi8q7sGI5FViG5Gm1D+Ic5/9FEg8iwsjCUpInr8ZlX6vQNerh5jupvc8n2\r\ntjCckcaYlEajhXfxiza5VGyDvA19glDLp/gsrDmdy3EbEbYhrnLJZ/Hq7sxqOKmA\r\nB3d0R/gBAoGBAOnW6Lfq8gkX0Wux2RbYRtCWIwwTQktR8k1Xqa2z+qJZC5Pclz6o\r\ncDL3Ebqr5XsobOZTEffLmutF6PJeUMt9QKTCJy34G2Ua7HldTdwnQIdwFrmx2REQ\r\nxpJr0qs79H91MYA2KE8D7m9kxT/AvtLHs3uqn5U3twBhICVeJIfbEBSVAoGBAMDv\r\nmJdmCZ7a4vT208xcpO9jqRM5ssFwKc1QdliJ+K8A5rkhKK4EZhKectSxFWBAMZbH\r\nj/Gg3mCpdnRP8+HtjJlCj++roGJKWe7EtLwk6xND/sDvkTCFT75sDkc/TrDxeKkg\r\n+D9lY3oO56S9cq5WEwiTF/wGRbLWeeWpxpkCbuiBAoGBAIj2maRv9ogg7Ua5CsB4\r\nuTGfXrPJ7ACV5OgsOnRFgL3qjf9zKxtGKUGQuU+vM2Bpky9s5I9RMFIxxnl/iK9H\r\nk79tAOnJ8rHLjUqgveP/EarCmmz1oICAZO6KPQCSw2iMSgwgrIdrdBLG4UmwKZTn\r\n51a9327VDVkpilyYwajZRJfpAoGBAIDeCC/nybk/PqScUO+JWelDyeS7fzhRMLqg\r\nacvNrDtIANPa4A8LhRoel1iVR0/Y0MaDWRPtfVtXFXWvCZj4UBb2LKnLGv9L5jHj\r\nVUzt484kdn5sBbBA+viXuBnCoU7ymsNs8CN4Fy+p0PNqTrU/bMsnZLfDgL5pH1ZV\r\nRf9hbq8BAoGAD8gHbkY3D87fkgoTRqdNh2+AUc7Y5aBsUkxhLC3o1ltTCZJ4Ly6c\r\n593+wvLvqI1iZUQ5cPXzMOp1mLtsD+K/o2vvJqiIvU6DU4cYfthknPGZ+1bSezVf\r\nSbXO3+G0vXtbDk1RjCs98Od0Q1Nzur0OUDU9+r0faIjrl36WsAvGJEA=\r\n-----END RSA PRIVATE KEY-----\r\n-----BEGIN CERTIFICATE-----\r\nMIIDwTCCAqmgAwIBAgIULPG5KdtzvUPbggZAjQwwc+wrQZswDQYJKoZIhvcNAQEL\r\nBQAwbzELMAkGA1UEBhMCSlAxDjAMBgNVBAgMBVRva3lvMRIwEAYDVQQHDAlNaW5h\r\ndG8ta3UxGDAWBgNVBAoMD05FQyBDb3Jwb3JhdGlvbjEMMAoGA1UECwwDUFhSMRQw\r\nEgYDVQQDDAsqLnB4ci5jby5qcDAgFw0yMDAzMjYxMjM3MDJaGA8yMTIwMDMwMjEy\r\nMzcwMlowbzELMAkGA1UEBhMCSlAxDjAMBgNVBAgMBVRva3lvMRIwEAYDVQQHDAlN\r\naW5hdG8ta3UxGDAWBgNVBAoMD05FQyBDb3Jwb3JhdGlvbjEMMAoGA1UECwwDUFhS\r\nMRQwEgYDVQQDDAsqLnB4ci5jby5qcDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCC\r\nAQoCggEBALA8CY8ZDsElrn2n0b0RFDtzC3si7jo9KV4dnKcOVTUt4Yn5b1pfHDJY\r\neYeKaeJQCOLvjC/u2CvIHOwZZOsuhmA1Ra0eCoIKTjyIu/Ge1Xrg+sNFruz0ubHJ\r\nY+4OC5NT8Z0uBtLiDgJvpr1HFKzWRTeIreYrN7JkhvTKemC4fuf1TB0+Px/Omqnd\r\n3VYCGk8/JHJ2zgBgQY459WLGbn+1LGz39zmzht/jaRP0ZNW/7N50ZDRXWceCuA/v\r\nXboyCX2epzZ+tUS9y6lJungLeqg6/9arojYjSzQ9a5NF/qIyfy2h7wTA98cUpuSD\r\nETIRGxDXIXCSwFbJ4w7hI4ZTGvfHZxUCAwEAAaNTMFEwHQYDVR0OBBYEFH1wO/lR\r\nGxFwgiP20dva7YZysZVDMB8GA1UdIwQYMBaAFH1wO/lRGxFwgiP20dva7YZysZVD\r\nMA8GA1UdEwEB/wQFMAMBAf8wDQYJKoZIhvcNAQELBQADggEBADL0jLXKzstpXMnn\r\nUhlr9AFxaA5n3sArc081n+HfYjY4vZwj5wnLRoy5xeYmHHEQ5lBwMHrS0MbsjMR2\r\njWIKcAhkD5KA6JrGoqqqUx30uToKVkWiwRjxxcMVOAsnBXa4V4Mvyi6sY/OTwic7\r\nIW3H6+UtcEXXpy199Q2GAhBg7Fb23+D8JaiDXG5hccB2J6hP9t8C2fWxKIokp2J4\r\n773pjj9Ke+ibjZ7GeaD4HYMMd30zVHJllRdVV7gpRCoyUKoNwkS722A5ZykUGldL\r\nv0DJpYdnexHNJWm5Bi8rubmeIIETI0VC8oYzYJCj1O3AK0DVkW2HPjpTxjuQZr7p\r\nl28YZGM=\r\n-----END CERTIFICATE-----\r\n"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/check/client": {
            "get": {
                "tags": [
                    "証明書確認"
                ],
                "description": "クライアント証明書の確認を確認する。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "sessionId",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "loginid",
                                "name": "test-user",
                                "mobilePhone": "0311112222",
                                "auth": {
                                    "add": true,
                                    "update": true,
                                    "delete": true
                                },
                                "lastLoginAt": "2020-01-01T00:00:00.000+0900",
                                "attributes": {},
                                "roles": [
                                    {
                                        "_value": 1,
                                        "_ver": 1
                                    }
                                ],
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "証明書の保存に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "レスポンス",
                                    "properties": {
                                        "isExist": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{serialNo}/{fingerPrint}": {
            "delete": {
                "tags": [
                    "証明書失効"
                ],
                "description": "シリアル番号、フィンガープリントを受け、失効します。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "sessionId",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "loginid",
                                "name": "test-user",
                                "mobilePhone": "0311112222",
                                "auth": {
                                    "add": true,
                                    "update": true,
                                    "delete": true
                                },
                                "lastLoginAt": "2020-01-01T00:00:00.000+0900",
                                "attributes": {},
                                "roles": [
                                    {
                                        "_value": 1,
                                        "_ver": 1
                                    }
                                ],
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    },
                    {
                        "name": "serialNo",
                        "description": "シリアル番号",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "example": "70F3C323BDEFDC1BFC9EBAA3CCE7EB257837BB61"
                    },
                    {
                        "name": "fingerPrint",
                        "description": "フィンガープリント",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "example": "85:EF:19:FE:99:97:3A:69:81:F5:28:ED:57:55:22:8A:59:AB:E5:5F"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "レスポンス",
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}